Replaced "== None" and "!= None" with "is None" and "is not None".
The previous constructs were not incorrect, but produced warnings from NumPy.
It gave a deprecation warning, saying that comparisons of arrays to None will
eventually change to an elementwise comparison, resulting in an array of
booleans instead of a single boolean.  This change will prevent GSim from
running into errors with future versions of NumPy.