Beginning conversion to GTK+3 and PyGObject.
Used PyGObject's pygi-convert.sh [1] to make the bulk of the conversion.  Did
some manual fixes after that to make it run.  It now has basic functionality
with GTK+3 through PyGObject (instead of the old PyGTK), and will run on both
Python 2 and Python 3.

A few things do not work as well as in the previous code and need fixing:
* The view of the GCode source is initially given zero width, requiring that
  the user manually pull over the divider in order to see it.  This should
  start at a position where the text can be seen.
* The origin of the graphics view is initially outside of the screen borders,
  requiring that the user manually pan to see it.  The view should start with
  the origin at the bottom left.
* The length scale is not drawn in the graphics view.  It should be.

[1] https://git.gnome.org/browse/pygobject/tree/pygi-convert.sh