Fixed drawing origin and scale issues with converted code.
Of the three issues identified in the quick conversion performed in the previous
commit, two have been fixed here.  The drawing origin and the length scale are
now both placed in the correct location.  The issue was caused by improperly
operating with the size of the toplevel window, when the size of the drawing
area needed to be used instead.