Fixed issue with text view width.
The final issue found in the quick conversion is fixed here.  The text view
would start with zero width, but now starts with the same width that it did in
the old GTK+2-based version.  The issue was that the text view was placed inside
a scroll box, which was placed in the adjustable pane.  The text view requested
a certain size, but this request was not being honoured.  Having the scroll box
make the request instead fixed the issue.

With this, the GTK+3-based program has feature-parity with the previous GTK+2-
based version.